python_basic_diploma/requirements.txt
aiogram==2.23.1
python-dotenv=1.0.1